Report package capabilities

This function summarises the capabilities of the package as compiled for the current machine, analogously to the base function capabilities(). It identifies the support available for various input and output formats.

divest.capabilities(what = NULL)

Arguments

what

A character vector of components to extract, or NULL, the default, which indicates the full set of capabilities.

Value

A named logical vector, indicating whether plain JPEG, JPEG-LS and JPEG2000 DICOM transfer syntaxes are supported by the current build of the package, and also whether zlib is available to produce compressed NIfTI output files from convertDicom().

See also

Author

Jon Clayden code@clayden.org

Examples

divest.capabilities()
#>     jpeg   jpegLS jpeg2000     zlib 
#>     TRUE    FALSE     TRUE     TRUE